requirements.txt
streamlit==1.4.0
Pillow>=8.1.1
nltk
contractions==0.0.48
inflect
BeautifulSoup4
textblob==0.15.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
spacy==2.3.5
gensim==3.8.3
pyLDAvis==2.1.2
seaborn==0.11.1
matplotlib==3.3.3
sklearn==0.0
scipy==1.5.4
numpy==1.20.0rc1
wordcloud==1.8.1
tqdm==4.54.1
audiomentations
textstat
opencv-python-headless
importlib-metadata==4.13.0
